requirements.txt
BTrees==4.2.0
cffi==1.4.2
cryptography==1.1.2
Flask==0.10.1
idna==2.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
ndg-httpsclient==0.4.0
persistent==4.1.1
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.4.2
requests==2.9.1
six==1.10.0
transaction==1.4.4
Werkzeug==0.11.3
zc.lockfile==1.1.0
ZConfig==3.1.0
zdaemon==4.1.0
ZEO==4.2.0b1
ZODB==4.2.0
zodbpickle==0.6.0
zope.interface==4.1.3
meld3==1.0.2
